loader apply style
