+++ examples/components/theme-configurator/utils/utils.js
@@ -70,13 +70,3 @@ export const getActionDisplayName = (key) => {
 export const getCategoryDisplayName = (key) => {
   return getNameFromI18N('category')[key] || key;
 };
-
-export const updateDomHeadStyle = (id, styleContent) => {
-  let styleTag = document.getElementById(id);
-  if (!styleTag) {
-    styleTag = document.createElement('style');
-    styleTag.setAttribute('id', id);
-    document.head.appendChild(styleTag);
-  }
-  styleTag.innerText = styleContent.replace(/@font-face{[^}]+}/, '');
-};
