+++ examples/components/theme/loader/api.js
